// clients/kpi-client/src/types/kPI_LyLich2C/kPI_LyLich2C.d.ts
export interface KPI_LyLich2CType extends EntityType {
  donViSuDungId: string;
	donViSuDungName?: string;
	phongBanId: string;
	phongBanName?: string;
	chucVuHienTai: string;
	chucVuHienTaiName?: string;
	maCanBo: string;
	userId?: string;
	userName?: string;
	roleCodes?: string[];
	roleNames?: string[];
	hoTen: string;
	avatar: string;
	gioiTinh: number;
	soHieuCCVC: string;
	ngayBoNhiem: Date;
	ngayBoNhiemLai: Date;
	ngaysinh: Date;
	email: string;
	status: string;
	tinhTrangHonNhan: string;
	tenKhac: string;
	phone: string;
	loaiHopDong: string;
	loaiHopDongName?: string;
	soCMND: string;
	ngayCapCMND: Date;
	noiCapCMND: string;
	noiSinhTinh: string;
	noiSinhXa: string;
	isNoiSinh: boolean;
	noiSinh: string;
	queQuanGoc: string;
	queQuanTinh: string;
	queQuanXa: string;
	queQuan: string;
	danToc: string;
	tonGiao: string;
	quocTich: string;
	hoKhauThuongTru_Tinh: string;
	hoKhauThuongTru_Xa: string;
	noiDangKyHKTT: string;
	noiOHienNay_Tinh: string;
	noiOHienNay_Xa: string;
	diaChiHienTai: string;
	coQuanTuyenDung: string;
	ngayTuyenDung: Date;
	nVaoCoQuanHienDangCongTac: Date;
	congViecChinh: string;
	soTruongCongTac: string;
	congViecLamLauNhat: string;
	chucDanh: string;
	chucDanhQuyHoach: string;
	loaiHinhDaoTao: string;
	boiDuongLanhDaoCapVu: string;
	quanLyNN: string;
	trinhDoGiaoDucPhoThong: string;
	trinhDoMax: string;
	trinhDoMaxName?: string;
	trinhDo: string;
	viTriViecLam: string;
	noteTrinhDoCM: string;
	lyLuanChinhTri: string;
	lyLuanChinhTriName?: string;
	quanLyNhaNuoc: string;
	quanLyNganh: string;
	tinHoc: string;
	thongTinTinHoc: string;
	tiengAnh: string;
	thongTinTiengAnh: string;
	ngoaiNgu: string;
	thongTinNgoaiNgu: string;
	tiengDanToc: string;
	thongTinTiengDanToc: string;
	ngayVaoDang: Date;
	ngayVaoDangChinhThucTxt: Date;
	noiKetNapDang: string;
	chucVuDangHienTai: string;
	chiBoSinhHoatDang: string;
	daiBieuHoiDongNhanDan: string;
	ngayVaoDoan: Date;
	noiKetNapDoan: string;
	chucVuDoan: string;
	ngayNhapNgu: Date;
	danhHieuPhongTang: string;
	danhHieuMax: string;
	hocHam: string;
	namPhongHocHam: Date;
	namPhongChucDanhKhoaHoc: Date;
	chuyenNganhHocHam: string;
	chucDanhKhoaHoc: string;
	levelThuongBinh: string;
	laConGiaDinhChinhSach: string;
	soBaoHiemXH: string;
	tPhanGiaDinhXuatThan: string;
	tPhanBanThanXuatThan: string;
	ngayThamGiaCachMang: Date;
	doiTuongChinhSach: string;
	luong: string;
	nguonThuKhac: string;
	nhanXetDanhGia: string;
	ngayKyQuyetDinh: Date;
	ngayHieuLuc: Date;
	lyDo: string;
	soQuyetDinh: string;
	nguoiKy: string;
	ngayHuongLuong: Date;
	maNgach: string;
	heSoLuong: number;
	bacLuong: number;
	ngachCongVienChuc: string;
	idBacLuong: string;
	loaiDieuChinhLuongLyLich: string;
	loaiLuong: string;
	soTienLuongThoaThuan: number;
	ngayBoNhiemChucDanh: Date;
	phanTramHuong: number;
	ngayHuongPhuCapThamNienVuotKhung: Date;
	vuotKhung: number;
}